requirements.txt
flask==2.3.3
flask-cors==4.0.0
pandas==2.1.0
numpy==1.26.4
praw==7.8.1
google-generativeai==0.7.2